JSON parser added

+++ src/Option.cpp
@@ -4,7 +4,7 @@ Option::Option() {}
 
 Option::Option(std::string information) {
     this->information = information;
-    this->choiceValue = (uint8_t) information[0];
+    this->choiceValue = (int) information[0];
 }
 
 Option::~Option() {}
@@ -13,7 +13,7 @@ std::string Option::getInformation() {
     return this->information;
 }
 
-uint8_t Option::getChoiceValue() {
+int Option::getChoiceValue() {
     return this->choiceValue;
 }
 
@@ -21,11 +21,11 @@ void Option::setInformation(std::string information) {
     this->information = information;
 }
 
-void Option::setChoiceValue(uint8_t choiceValue) {
+void Option::setChoiceValue(int choiceValue) {
     this->choiceValue = choiceValue;
 }
 
 void Option::assignOption(std::string information) {
     this->information = information;
-    this->choiceValue = (uint8_t) this->information[0];
+    this->choiceValue = (int) this->information[0];
 }
